+++ buildSrc/src/main/kotlin/ProjectMetadata.kt
@@ -2,6 +2,7 @@
 import io.github.rhdunn.gradle.maven.BuildType
 import io.github.rhdunn.gradle.maven.SupportedVariants
 import org.gradle.api.JavaVersion
+import org.jetbrains.kotlin.konan.target.KonanTarget
 
 /**
  * Information about the project.
@@ -52,6 +53,45 @@ object ProjectMetadata {
         val JvmTargets: List<JavaVersion> = listOf(
             JavaVersion.VERSION_11, // LTS
             JavaVersion.VERSION_17, // LTS
+        )
+
+        /**
+         * The default value for the `konan.variants` property.
+         */
+        val DefaultKonanVariants = SupportedVariants.All
+
+        /**
+         * Specifies which Java versions to target in Kotlin/JVM.
+         *
+         * @see <a href="https://kotlinlang.org/docs/native-target-support.html">Kotlin/Native target support</a>
+         */
+        val KonanTargets: List<KonanTarget> = listOf(
+            KonanTarget.ANDROID_ARM32, // Tier 3
+            KonanTarget.ANDROID_ARM64, // Tier 3
+            KonanTarget.ANDROID_X64, // Tier 3
+            KonanTarget.ANDROID_X86, // Tier 3
+            KonanTarget.IOS_ARM32, // Deprecated, to be removed in 1.9.20
+            KonanTarget.IOS_ARM64, // Tier 2
+            KonanTarget.IOS_SIMULATOR_ARM64, // Tier 1
+            KonanTarget.IOS_X64, // Tier 1
+            KonanTarget.LINUX_ARM32_HFP, // Deprecated, to be removed in 1.9.20
+            KonanTarget.LINUX_ARM64, // Tier 2
+            KonanTarget.LINUX_MIPS32, // Deprecated, to be removed in 1.9.20
+            KonanTarget.LINUX_MIPSEL32, // Deprecated, to be removed in 1.9.20
+            KonanTarget.LINUX_X64, // Tier 1 ; native host
+            KonanTarget.MACOS_ARM64, // Tier 1 ; native host
+            KonanTarget.MACOS_X64, // Tier 1 ; native host
+            KonanTarget.MINGW_X64, // Tier 3 ; native host
+            KonanTarget.MINGW_X86, // Deprecated, to be removed in 1.9.20
+            KonanTarget.TVOS_ARM64, // Tier 2
+            KonanTarget.TVOS_SIMULATOR_ARM64, // Tier 2
+            KonanTarget.TVOS_X64, // Tier 2
+            KonanTarget.WASM32, // Deprecated, to be removed in 1.9.20
+            KonanTarget.WATCHOS_ARM32, // Tier 2
+            KonanTarget.WATCHOS_ARM64, // Tier 2
+            KonanTarget.WATCHOS_SIMULATOR_ARM64, // Tier 2
+            KonanTarget.WATCHOS_X64, // Tier 2
+            KonanTarget.WATCHOS_X86, // Deprecated, to be removed in 1.9.20
         )
     }